+++ __tests__/hostApi/injectedWeb3Provider.spec.ts
@@ -8,7 +8,7 @@ import {
 } from '@novasamatech/host-api';
 import type { ContainerHandlerOf } from '@novasamatech/host-container';
 import { createContainer } from '@novasamatech/host-container';
-import { createNonProductExtensionEnableFactory } from '@novasamatech/product-sdk';
+import { createLegacyExtensionEnableFactory } from '@novasamatech/product-sdk';
 
 import type { SignerResult } from '@polkadot/types/types';
 import { AccountId } from '@polkadot-api/substrate-bindings';
@@ -21,7 +21,7 @@ async function setup() {
   const container = createContainer(providers.host);
   const sdkTransport = createTransport(providers.sdk);
 
-  const enable = await createNonProductExtensionEnableFactory(sdkTransport);
+  const enable = await createLegacyExtensionEnableFactory(sdkTransport);
   assert(enable, 'Enable function should be available');
 
   const injected = await enable();
@@ -36,7 +36,7 @@ describe('Host API: injected web3 provider', () => {
 
     const { container, injected } = await setup();
 
-    container.handleGetNonProductAccounts((_, { ok }) => ok(mockAccounts));
+    container.handleGetLegacyAccounts((_, { ok }) => ok(mockAccounts));
 
     const injectedAccounts = await injected.accounts.get();
 
@@ -58,7 +58,7 @@ describe('Host API: injected web3 provider', () => {
       signature: '0x0001',
     };
 
-    container.handleSignPayloadWithNonProductAccount((params, { ok }) => {
+    container.handleSignPayloadWithLegacyAccount((params, { ok }) => {
       return ok({ ...signerResult, signedTransaction: params.payload.method });
     });
 
@@ -89,7 +89,7 @@ describe('Host API: injected web3 provider', () => {
       signature: '0x0001',
     };
 
-    container.handleSignRawWithNonProductAccount((params, { ok }) => {
+    container.handleSignRawWithLegacyAccount((params, { ok }) => {
       return ok({ ...signerResult, signedTransaction: params.payload.value as HexString });
     });
 
@@ -126,11 +126,11 @@ describe('Host API: injected web3 provider', () => {
       },
     };
 
-    const createTransaction = vitest.fn<
-      ContainerHandlerOf<typeof container.handleCreateTransactionWithNonProductAccount>
-    >((_, { ok }) => ok(response));
+    const createTransaction = vitest.fn<ContainerHandlerOf<typeof container.handleCreateTransactionWithLegacyAccount>>(
+      (_, { ok }) => ok(response),
+    );
 
-    container.handleCreateTransactionWithNonProductAccount(createTransaction);
+    container.handleCreateTransactionWithLegacyAccount(createTransaction);
 
     const result = await injected.signer.createTransaction?.(payload);
 
@@ -142,7 +142,7 @@ describe('Host API: injected web3 provider', () => {
     const { container, injected } = await setup();
     const error = new RequestCredentialsErr.Rejected();
 
-    container.handleGetNonProductAccounts((_, { err }) => err(error));
+    container.handleGetLegacyAccounts((_, { err }) => err(error));
 
     await expect(injected.accounts.get()).rejects.toEqual(error);
   });
@@ -151,7 +151,7 @@ describe('Host API: injected web3 provider', () => {
     const { container, injected } = await setup();
     const error = new SigningErr.Rejected();
 
-    container.handleSignPayloadWithNonProductAccount((_, { err }) => err(error));
+    container.handleSignPayloadWithLegacyAccount((_, { err }) => err(error));
 
     await expect(
       injected.signer.signPayload?.({
@@ -175,7 +175,7 @@ describe('Host API: injected web3 provider', () => {
     const { container, injected } = await setup();
     const error = new SigningErr.Rejected();
 
-    container.handleSignRawWithNonProductAccount((_, { err }) => err(error));
+    container.handleSignRawWithLegacyAccount((_, { err }) => err(error));
 
     await expect(
       injected.signer.signRaw?.({
@@ -210,7 +210,7 @@ describe('Host API: injected web3 provider', () => {
       },
     };
 
-    container.handleCreateTransactionWithNonProductAccount((_, { err }) => err(error));
+    container.handleCreateTransactionWithLegacyAccount((_, { err }) => err(error));
 
     await expect(injected.signer.createTransaction?.(payload)).rejects.toEqual(error);
   });